Change external table logic to create Accumulo tables and set table configuration.
The original design of external tables was to have the connector be hands-off as far as
creating and configuring tables.  In practice, this has caused management issues with
making sure the external tables are properly configured.  This commit makes the behavior
of external tables align more with the use case: we want a simple SQL view on top of Accumulo
tables and we don't want to drop the Accumulo tables along with the Presto data.