Add polling duration to cardinality cache retrieval.
The LoadingCache is not ordered and, as a result, some cached results (or a result achieved from Accumulo in
a short time) that have higher cardinalities are returned a few milliseconds before a significantly lower result.
Before, the cache would take the first result below the configured threshold.  This change adds a sleep for
a poll duration, adding 'waves' of result retrieval.  The results of any completed tasks are taken, and the
smallest cardinality, if below the threshold, is used while the other tasks complete.